change files : fix delay in ft_usleep

+++ philo/includes/philo.h
@@ -6,7 +6,7 @@
 /*   By: elias <zanotti.elias@gmail.com>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/19 16:32:51 by elias             #+#    #+#             */
-/*   Updated: 2023/03/07 15:13:27 by elias            ###   ########.fr       */
+/*   Updated: 2023/05/03 11:47:52 by elias            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -59,7 +59,7 @@ int			ft_argv_checker(char **argv);
 //	TIME
 long long	ft_diff(long long start, long long end);
 long long	ft_get_time(void);
-void		ft_usleep(size_t sleep);
+void		ft_usleep(int sleep);
 //	UTILS
 int			ft_atoi(const char *str);
 int			ft_error(int error_code);
